+++ AppboyProject/android/app/src/main/java/com/appboyproject/MainApplication.java
@@ -3,17 +3,14 @@ package com.appboyproject;
 import android.app.Application;
 import android.util.Log;
 
-import com.appboy.AppboyLifecycleCallbackListener;
-import com.appboy.reactbridge.AppboyReactPackage;
-import com.appboy.support.AppboyLogger;
+import com.braze.BrazeActivityLifecycleCallbackListener;
+import com.braze.support.BrazeLogger;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
 
-import java.util.Arrays;
 import java.util.List;
 
 public class MainApplication extends Application implements ReactApplication {
@@ -40,8 +37,8 @@ public class MainApplication extends Application implements ReactApplication {
   @Override
   public void onCreate() {
     super.onCreate();
-    registerActivityLifecycleCallbacks(new AppboyLifecycleCallbackListener());
-    AppboyLogger.setLogLevel(Log.VERBOSE);
+    registerActivityLifecycleCallbacks(new BrazeActivityLifecycleCallbackListener());
+    BrazeLogger.setLogLevel(Log.VERBOSE);
     SoLoader.init(this, /* native exopackage */ false);
   }
 }