+++ AppboyProject/ios/AppboyProject/AppDelegate.m
@@ -70,11 +70,7 @@
   [[Appboy sharedInstance] userNotificationCenter:center didReceiveNotificationResponse:response withCompletionHandler:completionHandler];
 }
 
-- (void) application:(UIApplication *)application didReceiveRemoteNotification:(NSDictionary *)userInfo {
-  [[Appboy sharedInstance] registerApplication:application didReceiveRemoteNotification:userInfo];
-}
-
-- (void) application:(UIApplication *)application didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken {
+- (void)application:(UIApplication *)application didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken {
   [[Appboy sharedInstance] registerDeviceToken:deviceToken];
 }