+++ iOS/AppboyReactBridge/AppboyReactBridge/AppboyReactBridge.m
@@ -82,8 +82,6 @@ RCT_EXPORT_METHOD(setSDKFlavor) {
 RCT_EXPORT_METHOD(getInitialUrl:(RCTResponseSenderBlock)callback) {
   if ([AppboyReactUtils sharedInstance].initialUrlString != nil) {
     [self reportResultWithCallback:callback andError:nil andResult:[AppboyReactUtils sharedInstance].initialUrlString];
-  } else {
-    [self reportResultWithCallback:callback andError:@"Initial URL string was nil." andResult:nil];
   }
 }